adding some more testing, and cleaning up files
